% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{NonparaTransport-package}
\alias{NonparaTransport}
\alias{NonparaTransport-package}
\title{NonparaTransport: Nonparanormal Transport for Multivariate Distributions}
\description{
Computational tools for multivariate distributional data (collections of multivariate distributions per data point) under the nonparanormal (Gaussian copula) modeling. Provides fast C++ computation of nonparanormal transport (NPT) metric and the pairwise NPT matrix. Supports Fréchet regression of multivariate distributional responses on Euclidean covariates, component-wise R-squared for goodness-of-fit measurements, permutation-based hypothesis testing, and bivariate density visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IrinaStatsLab/NonparaTransport}
  \item Report bugs at \url{https://github.com/IrinaStatsLab/NonparaTransport/issues}
}

}
\author{
\strong{Maintainer}: Junyoung Park \email{junyoup@umich.edu}

Authors:
\itemize{
  \item Junyoung Park \email{junyoup@umich.edu}
  \item Irina Gaynanova \email{irinagn@umich.edu}
  \item Edward Shao \email{eshao@umich.edu}
}

}
\keyword{internal}
